Move Logprobs to separate file to fix circular import
